% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_na.R
\name{add_na}
\alias{add_na}
\title{Adds episodes without input variables as NA}
\usage{
add_na(x, ...)
}
\arguments{
\item{processed_table}{the processed table from process_dbl}

\item{reference_table}{the reference table}
}
\value{
A list with the original processed data in position 1 and missing
  data in 2
}
\description{
Returns a list with 2 tibbles of the exact same form as produced by
\code{\link{process_dbl}} NAs can be found in position 2 ("empty"). This is
useful in understanding how many episodes do not report on a certain field.
As both lists are in the exact same form, then can be bound easily.
}
