% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_locations.R
\name{attach_locations}
\alias{attach_locations}
\title{Add movement data to extracted table}
\usage{
attach_locations(connection, ltb)
}
\arguments{
\item{connection}{a EMAP database connection}

\item{ltb}{a time series table produced from \code{\link{extract}}}
}
\value{
a time series table with location data appendewd
}
\description{
Add movement data to extracted table
}
