% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.date_1d}
\alias{autoplot.date_1d}
\title{Plot Datetime}
\usage{
\method{autoplot}{date_1d}(x, sites_col = NULL, verify = TRUE)
}
\arguments{
\item{x}{a varified data frame}

\item{sites_col}{the HEX colour pallet for each site}
}
\description{
Plot datetime values from CC-HIC.
}
