% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.heat_cal}
\alias{autoplot.heat_cal}
\title{Plot Calendar Heatmap}
\usage{
\method{autoplot}{heat_cal}(object, ...)
}
\arguments{
\item{object}{a heat_cal object}

\item{...}{other arguments to be passed}
}
\description{
Builds the ggplot layers required to correctly display the calendar heatmap
}
