% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.real_1d}
\alias{autoplot.real_1d}
\title{Plot Real Valued Dataitem}
\usage{
\method{autoplot}{real_1d}(x, sites_col = NULL, verify = TRUE)
}
\arguments{
\item{x}{a varified data frame}

\item{sites_col}{the HEX colour pallet for each site}

\item{verify}{\code{logical} if use only verified data}
}
\description{
Plot real valued dataitem from CC-HIC.
}
