% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.real_2d}
\alias{autoplot.real_2d}
\title{Plot Real}
\usage{
\method{autoplot}{real_2d}(x, sites_col = NULL, verify = TRUE)
}
\arguments{
\item{x}{a varified data frame}

\item{sites_col}{the HEX colour pallet for each site}
}
\description{
Plot real values from CC-HIC
}
