% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.string_2d}
\alias{autoplot.string_2d}
\title{Plot String}
\usage{
\method{autoplot}{string_2d}(x, sites_col = NULL, verify = TRUE)
}
\arguments{
\item{x}{a varified data frame}

\item{sites_col}{the HEX colour pallet for each site}

\item{verify}{\code{logical} if use only verified data}
}
\description{
Plot String values from CC-HIC.
}
