% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.time_1d}
\alias{autoplot.time_1d}
\title{Plot Time}
\usage{
\method{autoplot}{time_1d}(x, sites_col = NULL, verify = TRUE)
}
\arguments{
\item{x}{a varified data frame}

\item{sites_col}{the HEX colour pallet for each site}

\item{verify}{\code{logical} if use only verified data}
}
\description{
Plot time values from CC-HIC
}
