% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterise_episodes.R
\name{characterise_episodes}
\alias{characterise_episodes}
\title{Characterise episodes}
\usage{
characterise_episodes(connection = NULL, .debug = FALSE)
}
\arguments{
\item{connection}{a connection to the CC-HIC database}

\item{.debug}{logical flag. If \code{TRUE} the function will extract from
internal package test data}
}
\value{
a tibble that characterises each episode. The attribute
\code{invalid_records} contains information related to invalid records and the
reason for invalidation
}
\description{
Characterise episodes
}
\details{
There are several ways to potentially define an episode.
Fundamentally, a start and an end datetime are required for an unambiguous
episode. The start datetime is already required to import episodes into the
CC-HIC database (it is a primary key) and so is ensured to be 100\% complete
if the schema definition has been applied. This function therefore serves
to identify the end of an episode, which is often much less clearly
defined. Potential sources of information, in order of precedence include:
\itemize{
\item datetime of unit departure: \code{0412}
\item date and time of in-unit death: \code{0042} and \code{0043}
\item date and time of brainstem death: \code{0044} and \code{0045}
\item date and time of body removed: \code{0038} and \code{0039}
}

Further, these end times must correlate with an outcome status:
\itemize{
\item vital status at discharge from ICU: \code{0097}
\item declaration of brainstem death: \code{0400}
}
}
