% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean Table}
\usage{
clean(tbl = NULL, col_names)
}
\arguments{
\item{tbl}{a table extracted from the CC-HIC database}

\item{col_names}{a vector of the names you want to change}

\item{dq_ref}{a tibble if you want to override the default boundary conditions
with the following column headings:
code_names
min_range
max_range}
}
\description{
This function takes an extracted table from either
\code{\link{extract_demographics}} or \code{\link{extract_timevarying}} and
returns the table with values that are out of range or impossible (as defined
in the data quality reference table)
}
