% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_column}
\alias{clean_column}
\title{Clean Column}
\usage{
clean_column(df, col_name, action = "NA")
}
\arguments{
\item{df}{table to clean}

\item{col_name}{column name to clean}

\item{action}{action of the cleaner}
}
\description{
Clean Column
}
