% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{combine}
\alias{combine}
\title{Combine extracted dataitems}
\usage{
combine(a, b)
}
\arguments{
\item{a}{an extracted data item returned by \code{\link[=extract]{extract()}}}

\item{b}{an extracted data item returned by \code{\link[=extract]{extract()}}}
}
\value{
a tibble with combined data
}
\description{
Combine extracted dataitems
}
\details{
Safely combines extracted dataitems in preparation for using
\code{\link[=evaluate_comparison]{evaluate_comparison()}}. If the data classes of the primary value differ
between the tables, then their missingness patterns are returned instead.
}
