% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_events.R
\name{coverage}
\alias{coverage}
\title{verify Event Plausibility - Coverage Checks (S3 Generic)}
\usage{
coverage(x, reference_tbl)
}
\arguments{
\item{x}{an extracted dataitem}

\item{reference_tbl}{reference table from \code{\link{make_reference}}}
}
\value{
a table describing the coverage over a defined time window
}
\description{
Checks to ensure that long term data item contribution is consistent. This is
because often back end changes occur in hospitals that silently disrupt the
ETL process, and as such, some dataitems disappear without warning.
}
