% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_events.R
\name{coverage_generic}
\alias{coverage_generic}
\title{Check Coverage}
\usage{
coverage_generic(x, reference_tbl, threshold = 10)
}
\arguments{
\item{x}{verified table}

\item{reference_tbl}{reference table}

\item{threshold}{threshold number of days in a month over which an anomaly is declared}
}
\description{
Check Coverage
}
