% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_events.R
\name{create_failure_log}
\alias{create_failure_log}
\title{Create a failure log}
\usage{
create_failure_log(failures, eval_code, description)
}
\arguments{
\item{failures}{a table (possible with nrow == 0 if there are no errors) with
the following columns: \code{site}, \code{episode_id}, \code{event_id}, \code{code_name},
\code{value}. If not already done so, the \code{value} column will be converted to
a string representation.}

\item{eval_code}{character vector length 1 with the data quality evaluation
code}

\item{description}{character vector length 1 with a standardised text
description of the code in use}
}
\value{

}
\description{
Create a failure log
}
\details{
Creates a failure log ready to be imported into the \code{events_quality}
table. This ensures there is a standardised format in use and returns an
empty table should there not be any failures.
}
