% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_heat.R
\name{create_grid}
\alias{create_grid}
\title{Calendar Gridlines}
\usage{
create_grid(calendar_template = NULL)
}
\arguments{
\item{calendar_template}{a calendar template from
\code{\link{create_calendar_template}}}
}
\value{
a grid for assembling a ggheatcal
}
\description{
This calulates the correct gridlines that need to be drawn when formatting
the calendar.
}
