% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{cross_entropy}
\alias{cross_entropy}
\title{Calculate Cross Entropy}
\usage{
cross_entropy(x)
}
\arguments{
\item{u}{a vector}

\item{v}{a vector the same length as \code{u}}
}
\value{
the cross entropy of \code{u} and \code{v}
}
\description{
IN DEVELOPMENT
}
\details{
Calculates the cross entropy of two vectors: \code{u} and \code{v}. This is
useful is decribing how different two distributions (particularly if
categorical) are.
}
