% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{daily_admssions}
\alias{daily_admssions}
\title{Daily Admissions for each Site}
\usage{
daily_admssions(reference = NULL, by_site = NULL)
}
\arguments{
\item{reference}{the reference table generated by \code{\link{make_reference}}}

\item{by_site}{the named site of interest as a character}
}
\value{
a tibble with the number of unique episodes admitted for a given day
}
\description{
Calculates the number of admissions for each calendar day, stratified
by site. This is a complete table, i.e. days with 0 admissions are not
listed
}
\examples{
daily_admssions(unique_cases, by_site == "UCL")
}
