% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{daily_events}
\alias{daily_events}
\title{Daily Events for HIC Data item by Site}
\usage{
daily_events(df = NULL, reference = NULL, by_site = NULL)
}
\arguments{
\item{df}{extracted data item}

\item{reference}{the reference table generated by \code{\link{make_reference}}}

\item{by_site}{the named site of interest as a character}
}
\value{
a tibble with the number of unique episodes admitted for a given day
}
\description{
Calculates the number of events contributed for each calendar day, stratified
by site. This is a complete table, i.e. days with 0 admissions are not
listed
}
