% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_chronology.R
\name{decompose_chronology}
\alias{decompose_chronology}
\title{Decompose chronology}
\usage{
decompose_chronology(connection, x, .debug = FALSE)
}
\arguments{
\item{connection}{a connection to the CC-HIC database.}

\item{x}{the output returned from the \code{\link[=evaluate_chronology]{evaluate_chronology()}} function.}

\item{.debug}{logical flag to run in debug mode.}
}
\value{
a tibble with (potential) error codes ready for import into the
  `events_quality` table.
}
\description{
Decompose chronology
}
\details{
Takes the failure logs from \code{\link[=evaluate_chronology]{evaluate_chronology()}} and decomposes it into
an error log ready for importing into the \code{events_quality} database table.
This is done because it was convenient to have an intermediate stage of this
process for plotting purposes.
}
\seealso{
\code{\link[=evaluate_chronology]{evaluate_chronology()}}
}
