% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{epi_length}
\alias{epi_length}
\title{Calculate Episode Length}
\usage{
epi_length(core_table, reference_table = NULL, events_table = NULL)
}
\arguments{
\item{core}{core table from \code{make_core()}}

\item{useDeath}{logical value to determine if this calculation should include death data}
}
\value{
a tibble with temporal episode figures, length of stay and a validation code
}
\description{
There are several ways to potentially define the end of an episode.
This algorithm performs a stepwise search to define the episode
end from:
\itemize{
  \item episode end date time: 0412
  \item date and time of in unit death: 0042 and 0043
  \item date and time of brainstem death: 0044 and 0045
  \item date and time of body removed: 0038 and 0039
}
}
\details{
If date and time of death preceed the discharge time (which may
in principle only be administrative) we overright the episode
end datetime with the death details

Validation is coded as follows:
\itemize{
  \item 0: episode validates as ok
  \item 1: there is no episode end datetime
  \item 2: episode stay was <= 0 - i.e. not possible
}
}
\examples{
epi_length(reference, tbls[["events"]])
}
