% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{episode_end_generic}
\alias{episode_end_generic}
\title{Collect Episode End Datetime}
\usage{
episode_end_generic(reference_table = NULL, events_table = NULL,
  code_name = "NIHR_HIC_ICU_0412")
}
\arguments{
\item{core_table}{core table from \code{make_core()}}
}
\value{
a tibble with mandatory episode defining characteristics.
}
\description{
Collects episodes end datetime for further processing
}
\examples{
episode_end_generic(reference, tbl[["events"]])
}
