% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{episode_validity}
\alias{episode_validity}
\title{Summarise Invalid Episodes}
\usage{
episode_validity(df)
}
\arguments{
\item{df}{the episode table returned from \code{\link{characterise_episodes}}}
}
\value{
a tibble containing summary information for validation at episode
  level
}
\description{
Provides an overview of the reasons for episode invalidation
}
\examples{
episode_validity(episode_length)
}
