% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterise_episodes.R
\name{episode_varacity}
\alias{episode_varacity}
\title{Summarise Non-Verifiable Episodes}
\usage{
episode_varacity(df)
}
\arguments{
\item{df}{the episode table returned from \code{\link{characterise_episodes}}}
}
\value{
a tibble containing summary information for validation at episode
  level
}
\description{
Provides an overview of the reasons for episode invalidation
}
