% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_bounds.R
\name{evaluate_bounds}
\alias{evaluate_bounds}
\title{Evaluate event temporal boundaries}
\usage{
evaluate_bounds(x = NULL, los_table = NULL)
}
\arguments{
\item{x}{extracted dataitem returned by [extract()]}

\item{los_table}{an episode length table returned by [characterise_episodes()]}
}
\value{
a tibble with (potential) error codes ready for import into the
  `events_quality` table.
}
\description{
Evaluate event temporal boundaries
}
\details{
Evaluates events for being outside the time boundaries of an
associated episode. This is not necessarily a problem, for example,
microbiology data could be returned after the patient has died or left the
ICU. However, it is often demonstrative of a bigger problem in the way in
which data has been exported from the source hospital. For example, data is
sometimes contributed from more than one episode, and duplicated across the
episodes.
}
