% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_chronology.R
\name{evaluate_chronology}
\alias{evaluate_chronology}
\title{Evaluate the chronology of key events}
\usage{
evaluate_chronology(connection = NULL, decompose = TRUE, .debug = FALSE)
}
\arguments{
\item{connection}{a connection to the CC-HIC database}

\item{decompose}{logical flag to output an object (if \code{TRUE}) ready to import
into the \code{events_quality} database table.}

\item{.debug}{logical flag to run in debug mode.}
}
\value{
if \code{decompose = FALSE} then the violations themselves are returned
for further processing. If \code{decompose = TRUE} (the default) then a tibble
with (potential) error codes ready for import into the \code{events_quality}
table.
}
\description{
Evaluate the chronology of key events
}
\details{
Evaluates the temporal sequence of key events with respect to the
ICNARC temporal schema. This function behaves differently to the other
\code{evaluate_} functions in that is accepts a database connection (not an
extracted dataitem). This is because the temporal schema is based upon many
events and so it was more convenient to extract them simultaneously with
\code{\link[wranglEHR:extract_demographics]{wranglEHR::extract_demographics()}}.
}
\seealso{
\code{\link[=decompose_chronology]{decompose_chronology()}}
}
