% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_comparison.R
\name{evaluate_comparison}
\alias{evaluate_comparison}
\title{Evaluate a comparison between events}
\usage{
evaluate_comparison(
  event_1 = NULL,
  event_2 = NULL,
  relationship = c("exists", "not_exists", "==", "!=", ">", ">=", "<", "<=")
)
}
\arguments{
\item{event_1}{an extracted event returned by \code{\link[=extract]{extract()}}. This event will
be placed on the \emph{left} hand side of the operator defined by the
\code{relationship} argument.}

\item{event_2}{an extracted event returned by \code{\link[=extract]{extract()}}. This event will
be placed on the \emph{right} hand side of the operator defined by the
\code{relationship} argument.}

\item{relationship}{character vector length 1 describing the type of
relationship to assert:
\itemize{
\item exists: both events exist at the same time (2d) or for the same
episode(1d)
\item not_exists: both events do not exist at the same time (2d) or for the
same episode(1d)
\item "==": event_1 == event_2
\item "!=": event_1 != event_2
\item ">": event_1 > event_2
\item ">=": event_1 >= event_2
\item "<": event_1 < event_2
\item "<=": event_1 <= event_2
}}
}
\value{

}
\description{
Evaluate a comparison between events
}
\details{
Checks the relationship between two events of interest. This takes
two forms depending upon whether or not the data items are of type 1d or 2d.
In all cases, the function evaluates the functional relationship of \code{event_1}
\code{test_relationship} \code{event_2}, where \code{test_relationship} acts as an operator
between the two.
\subsection{1-dimensional events (time invariant)}{

These items are checked without reference to any timepoint.
}

\subsection{2-dimensional events (time series)}{

These items are checked with respect to their timestamp, which is used to
identify co-incident events that are suitable for comparison.
\subsection{Examples of comparisons}{
\itemize{
\item systolic blood pressure is always above diastolic
\item mean blood pressure is always below systolic
\item all date time pairings exist for date and time of:
\itemize{
\item death
\item withdrawal
\item ready for discharge
\item brain stem death
}
}
}

}
}
\seealso{
\code{\link[=extract]{extract()}} \code{\link[=perform_evaluation]{perform_evaluation()}}
}
