% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_distribution.R
\name{evaluate_distribution}
\alias{evaluate_distribution}
\title{Evaluate distribution}
\usage{
evaluate_distribution(x = NULL, col_name = "value", threshold = 0.5)
}
\arguments{
\item{x}{extracted dataitem returned by [extract()]}

\item{col_name}{column name for the column of interest (defaults to \code{value})}

\item{threshold}{the limit value of the KS distance to consider a "different"
distribution. Default = 0.5}
}
\value{
a tibble with (potential) error codes ready for import into the
  `events_quality` table.
}
\description{
Evaluate distribution
}
\details{
Evaluate the distribution between numeric data using the KS distance
}
\seealso{
\code{\link[=ks_test]{ks_test()}}
}
