% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_duplication.R
\name{evaluate_duplicate}
\alias{evaluate_duplicate}
\title{Evaluate event duplication}
\usage{
evaluate_duplicate(x = NULL)
}
\arguments{
\item{x}{extracted dataitem returned by [extract()]}
}
\value{
a tibble with (potential) error codes ready for import into the
  `events_quality` table.
}
\description{
Evaluate event duplication
}
\details{
Evaluates events as being suspicious for being duplicated or not.
This is trained on the datetime of an object (if 2d), and looks for two
events that are perfectly co-incident. 1d events are checked for raw
duplication, since they should only occur once per episode anyway. Note, when
2 values are deemed to be duplicates, the one that first appears in the
database is cleared as verified, while the second (or more) is/are not.
}
