% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_episodes.R
\name{evaluate_episodes}
\alias{evaluate_episodes}
\title{Evaluate episodes}
\usage{
evaluate_episodes(episode_length = NULL, threshold = 10)
}
\arguments{
\item{episode_length}{a table returned by \code{\link[=characterise_episodes]{characterise_episodes()}}}

\item{threshold}{numeric scalar threshold number of days to use in calculation}
}
\value{
a table of the same form as \code{\link[=episode_length]{episode_length()}} but with failing
episodes removed and placed into the \code{invalid_records} attribute
}
\description{
Evaluate episodes
}
\details{
Determines which months are invalid based on a low contribution of data.
This is based upon the long term daily average for admissions. Days that
fall below 2 SD of the long term mean are flagged. If more than the
\code{threshold} number of flagged days occur in a single month, then the
month is invalidated and removed from further analysis.

This procedure removes episodes that occur during particularly sparse periods
(as this is likely that these months are contributing poor data) and return
only episodes that have a logical consistency in non-sparse months. The
analyst should consider if the denominator for the number of study months
should be changed following the use of this function.
}
