% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_events.R
\name{evaluate_events}
\alias{evaluate_events}
\title{Apply all quality evaluations to CC-HIC events}
\usage{
evaluate_events(x = NULL, los_table = NULL, reference_tbl = NULL, range = NULL)
}
\arguments{
\item{x}{extracted dataitem returned by [extract()]}

\item{los_table}{an episode length table returned by [characterise_episodes()]}

\item{range}{character string denoting the reference range to be used in
overriding the default reference range. A string is used as the form "[0,
5]" and "(0, 5)" are used to denote the type of limits to apply.}
}
\value{
a tibble with (potential) error codes ready for import into the
  `events_quality` table.
}
\description{
Apply all quality evaluations to CC-HIC events
}
\details{
Applies appropriate data quality evaluation flags to an extracted
dataitem. The evaluation functions that operate at the event level are
designed to perform a single simple check on a standardised object (the
table returned by \code{\link[=extract]{extract()}}) and return a standardised table
detailing the results of that check. Each check can be run independently if
so desired.
}
