% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_missingness.R
\name{evaluate_local_missingness}
\alias{evaluate_local_missingness}
\title{Evaluate (local) missingness}
\usage{
evaluate_local_missingness(x = NULL, reference_tbl = NULL)
}
\arguments{
\item{x}{extracted dataitem returned by [extract()]}

\item{reference_tbl}{reference table returned from \code{\link[=make_reference]{make_reference()}}}
}
\value{
a tibble with (potential) error codes ready for import into the
\code{events_missing} table.
}
\description{
Evaluate (local) missingness
}
\details{
Checks to ensure that long term data item contribution is
consistent. This evaluation is performed because sometimes data pipelines
in contributing hospitals change resulting in a silently disruption to data
contribution. As such, some events that were previously contributed
disappear without warning.
}
