% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_metadata.R
\name{evaluate_metadata}
\alias{evaluate_metadata}
\title{Evaluate metadata}
\usage{
evaluate_metadata(x)
}
\arguments{
\item{x}{extracted dataitem returned by [extract()]}
}
\value{
a tibble with (potential) error codes ready for import into the
  `events_quality` table.
}
\description{
Evaluate metadata
}
\details{
Checks to see if metadata is present. This does not (as yet) check
that the metadata is correct, only that it exsists.
}
