% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_origin.R
\name{evaluate_origin}
\alias{evaluate_origin}
\title{Evaluate event origins}
\usage{
evaluate_origin(x = NULL, los_table = NULL)
}
\arguments{
\item{x}{extracted dataitem returned by [extract()]}

\item{los_table}{an episode length table returned by [characterise_episodes()]}
}
\value{
a tibble with (potential) error codes ready for import into the
  `events_quality` table.
}
\description{
Evaluate event origins
}
\details{
evaluates that an event originates in a valid episode
}
