% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_ranges.R
\name{evaluate_range.date_1d}
\alias{evaluate_range.date_1d}
\title{Evaluate date event ranges}
\usage{
\method{evaluate_range}{date_1d}(x = NULL, ...)
}
\arguments{
\item{x}{extracted dataitem returned by [extract()]}
}
\description{
Evaluate date event ranges
}
\details{
Verifies that all dates are before now (i.e. not in the future) and not
before 1900-01-01, which seems reasonable.
}
