% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_ranges.R
\name{evaluate_range.datetime_1d}
\alias{evaluate_range.datetime_1d}
\title{Range Checks - Datetimes}
\usage{
\method{evaluate_range}{datetime_1d}(x = NULL, ...)
}
\arguments{
\item{x}{extracted dataitem returned by [extract()]}
}
\description{
Range Checks - Datetimes
}
\details{
Verifies that all dates are before now (i.e. not in the future) and not
before 1900-01-01 00:00:00, which seems reasonable.
}
