% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_ranges.R
\name{evaluate_range}
\alias{evaluate_range}
\title{Evaluate event ranges}
\usage{
evaluate_range(x = NULL, range = NULL)
}
\arguments{
\item{x}{extracted dataitem returned by [extract()]}

\item{range}{character vector length 1 denoting the reference range to be
used in overriding the default reference range. A string is used as the
form "\[0, 5\]" and "(0, 5)" are used to denote inclusive and exclusive
limits respectively.}
}
\value{
a tibble with (potential) error codes ready for import into the
  `events_quality` table.
}
\description{
Evaluate event ranges
}
\details{
Evaluates events as being out of range in the broadest possible
  sense. This includes: being numerically out of range, categorical data not
  in a predefined set, and codes not conforming to the appropriate standard
  (post codes, ICNARC etc). These ranges have been decided based upon ICNARC
  reference ranges, prior evidence (usually case report for exceptional
  values) or expert opinion. Reference ranges are all found in in the [qref]
  and can be overridden by supplying a range to the `range` argument.
}
