% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_ranges.R
\name{evaluate_range.time_1d}
\alias{evaluate_range.time_1d}
\title{Range Checks - Times}
\usage{
\method{evaluate_range}{time_1d}(x = NULL, ...)
}
\arguments{
\item{x}{extracted dataitem returned by [extract()]}
}
\description{
Range Checks - Times
}
\details{
Verifies that all times are between 00:00:00 and 23:59:59
}
