% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_ranges.R
\name{evaluate_range_numeric}
\alias{evaluate_range_numeric}
\title{Evaluate numeric event ranges}
\usage{
evaluate_range_numeric(x = NULL, range = NULL)
}
\arguments{
\item{x}{extracted dataitem returned by [extract()]}

\item{range}{character vector length 1 denoting the reference range to be
used in overriding the default reference range. A string is used as the
form "\[0, 5\]" and "(0, 5)" are used to denote inclusive and exclusive
limits respectively.}
}
\description{
Evaluate numeric event ranges
}
