% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_ranges.R
\name{evaluate_range_string}
\alias{evaluate_range_string}
\title{Evaluate string event ranges (set membership)}
\usage{
evaluate_range_string(x = NULL, range = NULL)
}
\arguments{
\item{x}{extracted dataitem returned by [extract()]}

\item{range}{character vector containing the valid and complete set from
which the data item values can take. If `NULL` (the default) the values
from `qref$possible_values` are taken.}
}
\description{
Evaluate string event ranges (set membership)
}
