% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{event_occurrances}
\alias{event_occurrances}
\title{Event Occurrances for each Site}
\usage{
event_occurrances(extracted_event = NULL, by_site = "UCL")
}
\arguments{
\item{extracted_event}{extracted HIC event}

\item{by_site}{a site code as a character vector}
}
\value{
a tibble with the number of unique episodes admitted for a given day
}
\description{
Calculates the number of event occurances for each calendar day, stratified
by site. Days with 0 event submissions are not listed
}
\examples{
event_occurrances(x, by_site == "UCL")
}
