% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_timevarying.R
\name{expand_missing}
\alias{expand_missing}
\title{Fill in 2d Table to make a Sparse Table}
\usage{
expand_missing(df, cadance = 1)
}
\arguments{
\item{df}{a dense time series table produced from extract_timevarying}

\item{cadance}{the cadance by which you want to expand the table (default = 1 hour)}
}
\value{
a sparse time series table
}
\description{
The extract_timevarying returns a non-sparse table (i.e. rows/hours with
no recorded information for a patient are not presented in the table)
This function serves to expand the table and fill missing rows with NAs.
This is useful when working with most time-series aware stats packages
that expect a regular cadance to the table.
}
\examples{
\\{dontrun
expand_missing(tb_1)
}
}
