% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract}
\alias{extract}
\title{Extract CC-HIC Data}
\usage{
extract(core_table = NULL, code_name = NULL, translate_site = NULL)
}
\arguments{
\item{core_table}{core table returned from \code{\link[=make_core]{make_core()}}}

\item{code_name}{a string vector of length 1 with the CC-HIC data item to be
extracted.}

\item{translate_site}{a lookup table to translate site names on export in a
consistent way. The table requires two columns: \code{site} and \code{translation}.
The \code{site} column should be a character vector with names of the existing
sites as they are stored in the database (e.g. \code{c("GSTT", "RGT")} etc).
The \code{translation} column should be a character vector with new names to
obfuscate the real names (e.g. \code{c("Site A", "Site B")} etc). This is a 1:1
mapping so doesn't confer any real privacy advantage, but is useful as a
means to casually obscure data origins.}
}
\value{
A tibble with 1 row per event
}
\description{
Extract CC-HIC Data
}
\details{
Extracts CC-HIC data from the database events tables and appends them with
the correct class for further processing.
}
\examples{
\dontrun{
# ctn <- DBI::dbConnect() create database connection
core <- make_core(ctn)
df <- extract(core, "NIHR_HIC_ICU_0108")
}
}
