% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_1d}
\alias{extract_1d}
\title{Extract 1d Values}
\usage{
extract_1d(core_table = NULL, input = NULL, data_location = NULL)
}
\arguments{
\item{core_table}{the core table from make_core}

\item{input}{the HIC code for the variable of interest}

\item{data_location}{the column name that stores the primary data for this
variable}
}
\value{
a tibble with HIC data for a specified variable
}
\description{
This function extracts the correct column from the CC-HIC database
depending upon what type of data is called for
}
\examples{
extract_1d(core, input = "NIHR_HIC_ICU_0409", data_location = "integer")
}
