% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_2d}
\alias{extract_2d}
\title{Extract 2d Values}
\usage{
extract_2d(core_table = NULL, input = NULL, data_location = NULL)
}
\arguments{
\item{core_table}{a core table}

\item{input}{the input variable of choice}

\item{data_location}{the column name that stores the primary data for this
variable}
}
\value{
a long table with 1 row per event from the CC-HIC database
}
\description{
This function extracts the correct column from the CC-HIC database
depending upon what type of data is called for. It additionally pulls
out the datetime column, which accompanies any data for this class
}
