% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_demographics.R
\name{extract_demographics}
\alias{extract_demographics}
\title{Extract 1d data from CC-HIC}
\usage{
extract_demographics(
  connection = NULL,
  episode_ids = NULL,
  code_names = NULL,
  rename = NULL
)
}
\arguments{
\item{connection}{a CC-HIC database connection}

\item{episode_ids}{an integer vector of episode ids from the CC-HIC DB
that you want to extact. The default (NULL) is to extract all.}

\item{code_names}{a character vector of CC-HIC codes}

\item{rename}{a character vector of names you want to relabel CC-HIC codes
as, or NULL (the default) if you do not want to relabel.}
}
\value{
A tibble of 1d data
}
\description{
Takes a remote database contection to CC-HIC, a vector of HIC codes (and
optionally a vector of labels to rename the codes) and returns a table with 1
row per patient and 1 column per data item.
}
\examples{
db_pth <- system.file("testdata/synthetic_db.sqlite3", package = "inspectEHR")
ctn <- connect(sqlite_file = db_pth)
hic_codes <- "NIHR_HIC_ICU_0409"
new_labels <- "apache_score"
extract_demographics(ctn, episode_ids = 13639:13643, hic_codes, new_labels)
DBI::dbDisconnect(ctn)
}
