% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_flags.R
\name{flag_all}
\alias{flag_all}
\title{Apply all validation flags to an extracted data item}
\usage{
flag_all(x, los_table = NULL)
}
\arguments{
\item{x}{extracted data item}

\item{los_table}{episode length table}
}
\value{
a tibble with flags applied
}
\description{
Applies basic validation flags to an extracted data item.
This is a wrapper function with conditional logic to flag for:
data item out of value range
data item out of temporal range of the episode
duplication of item
periodicity
}
\details{
periodicity checking is conditional on the preceding 3 flags,
as only those events that validate are taking into consideration
of periodicity checking
}
\examples{
hr <- extract(core, input = "NIHR_HIC_ICU_0108")
flag_all(hr, episode_length)
}
