% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_flags.R
\name{flag_bounds}
\alias{flag_bounds}
\title{Flag Events by Episode Boundaries (S3 Generic)}
\usage{
flag_bounds(x, los_table = NULL)
}
\arguments{
\item{x}{an extracted tibble}

\item{...}{}
}
\value{
a tibble of the same length as x with the following values
\describe{
  \item{-1}{event occurred significantly before ICU episode}
  \item{0}{event occurred during ICU episode}
  \item{+1}{event occurred significantly after ICU episode}
  \item{NA}{Foundation elements were not availible and so this parameter could not be calculated}
}
}
\description{
If the event of concern is a string, since only Airway and Organism are time
varying string fields the behavior on hic_str differs.
}
\examples{

flag_bounds(x, los_table)
}
