% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_flags.R
\name{flag_bounds_2d}
\alias{flag_bounds_2d}
\title{Flag boundaries of 2d data items}
\usage{
flag_bounds_2d(x = NULL, los_table = NULL)
}
\arguments{
\item{x}{an extracted nhic event table}

\item{los_table}{episode length table}
}
\value{
a two column tibble with internal id and duplication status
}
\description{
Sets boundary conditions for events in relation to episodes
}
\examples{
flag_bounds_2d(x, los_table)
}
