% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_flags.R
\name{flag_duplicate}
\alias{flag_duplicate}
\title{Flag Duplicates (S3 Generic)}
\usage{
flag_duplicate(x)
}
\arguments{
\item{x}{an extracted nhic event table}

\item{...}{}
}
\value{
a two column tibble with internal id and duplication status
}
\description{
Flags nhic events for duplication. Duplicate events are assigned the value 1L
}
\examples{
flag_duplicate(x)
}
