% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_flags.R
\name{flag_periodicity}
\alias{flag_periodicity}
\title{Flags for Periodicity of a Data Item (S3 Generic)}
\usage{
flag_periodicity(x, los_table = NULL)
}
\arguments{
\item{x}{an extracted nhic events table}

\item{los_table}{episode length table}
}
\value{
an extension of x with a mutated periodicity column
}
\description{
Takes a data item tibble that has been pre-flagged with range, boundary and duplicate errors
and provides a column indicating the periodicity of that data item stratified by each episode
number. This is a similar concept to the coverage mapping, although is looking at patient
level rather than site level.
}
\details{
These are only defined for 2d data

Periodicity is given as event count per los unit. The default behaviour will be events/day.
}
\examples{
flag_periodicity(x, los_table)
}
