% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_flags.R
\name{flag_periodicity_generic}
\alias{flag_periodicity_generic}
\title{Flag periodicity of 2d class}
\usage{
flag_periodicity_generic(x, los_table = NULL)
}
\arguments{
\item{x}{data item}

\item{los_table}{episode length table}
}
\value{
a mutated tibble from x including a periodicity column
}
\description{
Flag periodicity of 2d class
}
\examples{
flag_periodicity(x, los_table)
}
