% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_flags.R
\name{flag_range}
\alias{flag_range}
\title{Flag Events by Reference Range (S3 Generic)}
\usage{
flag_range(x = NULL)
}
\arguments{
\item{x}{an extracted event table}

\item{...}{further arguments to pass to the S3 method}
}
\value{
a tibble of the same length as x with the following features:
\describe{
  \item{-1}{event is below the defined range of plausibility}
  \item{0}{event is valid}
  \item{+1}{event is above the defined range of plausibility}
}
}
\description{
Flags each event as being in-range (0), out-of-range high (+1) or
out-of-range low (-1). These ranges have been calibrated by hand based on
prior evidence or expert opinion. Reference ranges are all found in
\code{\link{qref}}.
}
\examples{
flag_range(x)
}
