% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_codes.R
\name{generate_icnarc}
\alias{generate_icnarc}
\title{Generate ICNARC Diagnostic Codes}
\usage{
generate_icnarc(size = 1)
}
\arguments{
\item{size}{the number of codes you wish to generate}
}
\value{
a character vector of length \code{size} with ICNARC diagnostic codes
}
\description{
Generate ICNARC Diagnostic Codes
}
\examples{
generate_icnarc()
}
