% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_codes.R
\name{generate_nhs}
\alias{generate_nhs}
\title{Generate NHS Numbers}
\usage{
generate_nhs(size = 1)
}
\arguments{
\item{size}{the number of codes you wish to generate}
}
\value{
a character vector of length \code{size} with NHS Numbers
}
\description{
Generate random codes that fit the NHS number formatting specification.
This is useful when generating synthetic patients or testing the varification
functions.
}
\examples{
generate_nhs()
generate_nhs(5)
}
