% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_codes.R
\name{generate_post_code}
\alias{generate_post_code}
\title{Generate Post Code}
\usage{
generate_post_code(size = 1)
}
\arguments{
\item{size}{number of post codes you wish to generate}
}
\value{
a character vector of length \code{size} of post codes
}
\description{
Generate UK Post Codes. Note, this does not garuntee that the postcode
exists, only that is corresponds to the correct functional form of a UK
postcode.
}
\examples{
generate_post_code()
generate_post_code(5)
}
