% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarHeat.R, R/calendar_heat.R
\name{ggHeatCal_episodes}
\alias{ggHeatCal_episodes}
\title{Plot Calendar Heatmap - Admission episodes}
\usage{
ggHeatCal_episodes(x, gridLines, Title, max_limit = 25)

ggHeatCal_episodes(x, gridLines, Title, max_limit = 25)
}
\arguments{
\item{x}{admissions object}

\item{gridLines}{gridlines object}

\item{Title}{title of the final plot}

\item{max_limit}{the maximim limit of the scale to display}

\item{x}{admissions object}

\item{gridLines}{gridlines object}

\item{Title}{title of the final plot}

\item{max_limit}{the maximim limit of the scale to display}
}
\value{
a ggplot2 object

a ggplot2 object
}
\description{
Builds the ggplot layers required to correctly display the calendar heatmap

Builds the ggplot layers required to correctly display the calendar heatmap
}
\examples{
ggHeatCal_episodes(x, gridlines, "UCL Admission Heatmap")
ggHeatCal_episodes(x, gridlines, "UCL Admission Heatmap")
}
