% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarHeat.R, R/calendar_heat.R
\name{ggHeatCal_events}
\alias{ggHeatCal_events}
\title{Plot calendar heatmap for events}
\usage{
ggHeatCal_events(x, gridLines, Title, max_limit = 24)

ggHeatCal_events(x, gridLines, Title, max_limit = 24)
}
\arguments{
\item{x}{modified hic event from \code{\link{event_occurances}}}

\item{gridLines}{gridlines to be made by \code{\link{create_grid}}}

\item{Title}{title of the plot}

\item{max_limit}{the maximim limit of the scale to display}

\item{x}{modified hic event from \code{\link{event_occurances}}}

\item{gridLines}{gridlines to be made by \code{\link{create_grid}}}

\item{Title}{title of the plot}

\item{max_limit}{the maximim limit of the scale to display}
}
\value{
a ggplot2 object

a ggplot2 object
}
\description{
plots a cc-hic event over a calendar heatmap for inspection

plots a cc-hic event over a calendar heatmap for inspection
}
\examples{
ggHeatCal_events(df, gridlines, "Heart Rate Heatmap")
ggHeatCal_events(df, gridlines, "Heart Rate Heatmap")
}
