% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{identify_spells}
\alias{identify_spells}
\title{Characterise Spells}
\usage{
identify_spells(df = NULL, minutes = 30)
}
\arguments{
\item{df}{episode length table}

\item{minutes}{numeric scalar to define transition period}
}
\value{
a table with episodes reconciled as spells
}
\description{
Some sites have patients check out of one ICU and into another (for example
ICU stepdown to HDU). This checks to see if patients are discharged from one
unit and admitted to another wihtin a pre-defined time period, specified in
the minutes argument.
}
\details{
This only evaluates episodes that have already been flagged as valid by the
\code{\link{characterise_episodes}} function.
}
\examples{
identify_spells(episodes, episodes)
}
