% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-inspectEHR.R
\docType{package}
\name{inspectEHR}
\alias{inspectEHR}
\title{inspectEHR}
\description{
This package contains functions for data quality evaluation of the CC-HIC
database. This follows the framework described by Kahn and approaches the
problem through verification and validation. It's sister package
\code{wranglEHR} contains several functions to help extract and work with the
data.
}
