% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse Logistic Function}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{a numeric vector on the logistic scale}
}
\value{
a numeric vector on the probability scale
}
\description{
Inverse Logistic Function
}
\examples{
inv_logit(-4:4)
}
