% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lookup_hic}
\alias{lookup_hic}
\title{lookup CC-HIC codes}
\usage{
lookup_hic(search_term = "NIHR_HIC_ICU_0108", type = c("code", "name"))
}
\arguments{
\item{search_term}{code name for a CC-HIC data item, either as a full string
(e.g. "NIHR_HIC_ICU_0108") or as a number (e.g. 108), or the short_name
of the concept.}

\item{"code"}{or "name" depending if you want to search by hic_code or
short_name}
}
\value{
a lookup table
}
\description{
lookup CC-HIC codes
}
\examples{
lookup_hic(108)
lookup_hic("NIHR_HIC_ICU_0108")
}
