% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_base_calendar}
\alias{make_base_calendar}
\title{Make base calendar

Creates a base cadence calendar for all contributing sites so that each
year and month of submission is enumerated.}
\usage{
make_base_calendar(reference_tbl, resolution = c("day", "month"))
}
\arguments{
\item{reference_tbl}{a reference table returned from \code{\link[=make_reference]{make_reference()}}}

\item{resolution}{character vector length 1 with the desired base resolution
of the table}
}
\value{

}
\description{
Make base calendar

Creates a base cadence calendar for all contributing sites so that each
year and month of submission is enumerated.
}
