% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_core}
\alias{make_core}
\title{Make core table}
\usage{
make_core(connection = NULL, .debug = FALSE)
}
\arguments{
\item{connection}{a database connection returned by [DBI::dbConnect()]}

\item{.debug}{logical flag to use internal simulated data for testing}
}
\value{
a remote database table
}
\description{
Make core table
}
\details{
Prepares the appropriate table joins in the CC-HIC database to ready
  for querying
}
\examples{
\dontrun{
# Make a database connection object using DBI::dbConnect()
# ctn <- DBI::dbConnect()
core <- make_core(ctn)
head(core)
DBI::dbDisconnect(ctn)
}
}
