% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{make_dict}
\alias{make_dict}
\title{Produce CC-HIC data dictionary}
\usage{
make_dict(metadata = .variables)
}
\arguments{
\item{metadata}{CC-HIC variables table}
}
\value{
a data dictionary
}
\description{
Uses the variable table of the CC-HIC database to produce a lightweight data
dictionary
}
