% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_heat.R
\name{make_heatcal}
\alias{make_heatcal}
\title{Make a heat_cal object}
\usage{
make_heatcal(reference_tbl = NULL, dataitem_tbl = NULL, site = NULL,
  date_boundaries = NULL, max_limit = NULL)
}
\arguments{
\item{reference_tbl}{reference table produced from \code{\link{make_reference}}}

\item{dataitem_tbl}{an extracted dataitem from \code{\link{extract}}}

\item{site}{character string of hospital site to plot}

\item{date_boundaries}{character vector of format YYYY-MM-DD (or something
that is parsable by \code{lubridate::ymd}) defining the start and end
dates of the calendar. If nothing is supplied, this is derrived from the
data}

\item{max_limit}{the upper bound for determining colour scale. This is useful
if you have extreme outliers to enable good constrast over the range of
interest}
}
\value{
an object of class heat_cal
}
\description{
Make a heat_cal object
}
