% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report.R
\name{make_report}
\alias{make_report}
\title{Make HIC Report Data}
\usage{
make_report(database = "lenient_dev", username = "edward",
  password = "superdb", host = "localhost", port = 5432,
  sqlite_file = NULL, path_name = NULL, spell_boundary_mins = 60,
  write_plots = TRUE)
}
\arguments{
\item{database}{the database name for connection}

\item{username}{the database username}

\item{password}{the database password}

\item{host}{the database server host IP address}

\item{port}{the database port number}

\item{sqlite_file}{the filename for the sqlite database (if using it)}

\item{path_name}{the path name for output files}

\item{spell_boundary_mins}{the time gap you are willing to allow for the
definition of spells. This will automatiically check against your
contraint +/- 60 mins so you can see how sensitive your calcuation is.}

\item{write_plots}{true/false to output plots. Otherwise only data processing takes place,
which can be considerably quicker.}
}
\description{
This is the main function for processing data from CC-HIC. It wraps all
other functions to evaluate the whole database in one sweep.
This can take some time!
}
\details{
Specify the output folder in "path_name" when running this function,
path_name should be made availible with the following folders:
plots
working_data

report.rmd can be run after this function has successfully completed to
produce a full report for CC-HIC
}
\examples{
make_report(database = "test_db", username = "abcde", password = "qwerty", path_name = "~/outputs")
make_report(database = "test_db", sqlite_file = "~/database.sqlite", path_name = "~/outputs")
}
