% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_datetime}
\alias{merge_datetime}
\title{Merge datetime columns}
\usage{
merge_datetime(date, time, format = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{date}{a vector of class date, or string of default format \verb{\%Y-\%m-\%d}}

\item{time}{a vector of class time, or string of default format \verb{\%H:\%M:\%S}}

\item{format}{a string vector of length 1 with the format to parse the
combined date and time information. See \code{\link[base:strptime]{base::strptime()}}.}
}
\description{
Merge datetime columns
}
\details{
Dates and times are often split across concepts in CC-HIC. This function
takes a date and time column vectors and returns the combined asset as a
datetime class.
}
\seealso{
\code{\link[base:strptime]{base::strptime()}}
}
