% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{missingness}
\alias{missingness}
\title{Check for 1d Missingness}
\usage{
missingness(x, reference)
}
\arguments{
\item{x}{}
}
\value{
a tibble with the following columns:
\describe{
  \item{event}{name of the CC-HIC event}
  \item{site}{name of the BRC from which the event originates}
  \item{count}{total count of validated events recieved}
  \item{total}{total popululation}
  \item{missingness}{1 - count/total}
}
}
\description{
Check for 1d Missingness
}
\examples{
missingness(df)
}
