% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_evaluation.R
\name{perform_evaluation}
\alias{perform_evaluation}
\title{Perform CC-HIC Data Quality Evaluation}
\usage{
perform_evaluation(
  connection = NULL,
  output_folder = NULL,
  translate_site = NULL,
  test_comparisons = NULL,
  verbose = TRUE,
  .debug = FALSE
)
}
\arguments{
\item{output_folder}{character vector length 1 for the file path name for
plots to be stored. If left as \code{NULL} (the default) then no plots will be
written out.}

\item{translate_site}{a lookup table to translate site names on export in a
consistent way. The table requires two columns: \code{site} and \code{translation}.
The \code{site} column should be a character vector with names of the existing
sites as they are stored in the database (e.g. \code{c("GSTT", "RGT")} etc).
The \code{translation} column should be a character vector with new names to
obfuscate the real names (e.g. \code{c("Site A", "Site B")} etc). This is a 1:1
mapping so doesn't confer any real privacy advantage, but is useful as a
means to casually obscure data origins.}

\item{test_comparisons}{a lookup table with the following columns:
\itemize{
\item event_a (chr): \code{code_name} for event to be used on the left hand side
of the evaluation in \code{\link[=evaluate_comparison]{evaluate_comparison()}}.
\item event_b (chr): \code{code_name} for event to be used on the right hand side
of the evaluation in \code{\link[=evaluate_comparison]{evaluate_comparison()}}.
\item operation (chr): string to identify the comparison to be made. See the
documentation in the \code{\link[=evaluate_comparison]{evaluate_comparison()}} function.
}

The default behaviour is to evaluate the comparisons found in the
\code{.comparisons_lookup} object. This object also serves to illustrate the
correct format to use.}

\item{verbose}{logical flag to print progress to the console}

\item{.debug}{logical flag to use internal test data}

\item{database}{a database object returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}}
}
\value{
TRUE if completes without errors
}
\description{
Perform CC-HIC Data Quality Evaluation
}
\details{
Performs a full data quality evaluation for CC-HIC. This wraps all
other functions in \code{inspectEHR} as necessary to evaluate the whole CC-HIC
database in one easy sweep. This can take some time so please grab a
coffee!
}
