% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_contribution}
\alias{plot_contribution}
\title{Plot CC-HIC event contribution}
\usage{
plot_contribution(x = NULL, reference_tbl = NULL, sites_col = NULL)
}
\arguments{
\item{x}{extracted dataitem returned by [extract()]}

\item{reference_tbl}{reference table returned from [make_reference()]}

\item{sites_col}{character vector containing HEX codes for the colour used to
represent each site. The names of the vector should be the site names as
they are to be displayed in the plot legend.}
}
\description{
Plot CC-HIC event contribution
}
