% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_datetime}
\alias{plot_datetime}
\title{Plot Datetime}
\usage{
plot_datetime(x, code_name, all_sites.col)
}
\arguments{
\item{x}{an extracted (using \link{\code{extract}}) and flagged (using \link{\code{flag_all}}) table}

\item{code_name}{the code name of interest}

\item{all_sites.col}{the HEX colour pallet for each site}
}
\description{
Plot datetimes from CC-HIC
}
