% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_default}
\alias{plot_default}
\title{Default Plot}
\usage{
plot_default(x, code_name, all_sites.col)
}
\arguments{
\item{x}{an extracted (using \link{\code{extract}}) and flagged (using \link{\code{flag_all}}) table}

\item{code_name}{the code name of interest}

\item{all_sites.col}{the HEX colour pallet for each site}
}
\value{
a default plot for a single hic code
}
\description{
Produce a default plot (eCDF or histogram) for a CC-HIC event code.
}
