% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ecdf}
\alias{plot_ecdf}
\title{Plot CC-HIC event eCDF}
\usage{
plot_ecdf(x, sites_col = NULL, col_name = c("value", "datetime"))
}
\arguments{
\item{x}{extracted dataitem returned by [extract()]}

\item{sites_col}{character vector containing HEX codes for the colour used to
represent each site. The names of the vector should be the site names as
they are to be displayed in the plot legend.}

\item{col_name}{the target column name to display. Defaults to "value".}
}
\description{
Plot CC-HIC event eCDF
}
