% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_full}
\alias{plot_full}
\title{Full Plot}
\usage{
plot_full(x, code_name, all_sites.col)
}
\arguments{
\item{x}{an extracted (\code{\link{extract}}) table}

\item{code_name}{the CC-HIC codename of interest.}
}
\description{
Plots the full distribution of a HIC event, regardless of any outliers or invalid data.
}
