% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heat_cal}
\alias{plot_heat_cal}
\title{Plot Calendar Heatmap}
\usage{
plot_heat_cal(heat_cal, scale_max = NULL)
}
\arguments{
\item{heat_cal}{a table of class `heat_cal` returned from
\code{\link{make_heatcal}}}

\item{scale_max}{maximum number of daily admissions to scale fill}
}
\description{
Builds the ggplot layers required to correctly display the calendar heatmap
}
