% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot CC-HIC event histogram

Plot a histogram of a CC-HIC event.}
\usage{
plot_hist(x, sites_col = NULL)
}
\arguments{
\item{x}{extracted dataitem returned by [extract()]}

\item{sites_col}{the HEX colour pallet for each site}
}
\description{
Plot CC-HIC event histogram

Plot a histogram of a CC-HIC event.
}
\concept{plotting}
