% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_missing_events}
\alias{plot_missing_events}
\title{Plot missing CC-HIC events}
\usage{
plot_missing_events(missing_events)
}
\arguments{
\item{missing_events}{a missing events table returned from
\code{\link[=evaluate_global_missingness]{evaluate_global_missingness()}}}
}
\description{
Plot missing CC-HIC events
}
