% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_periodicity}
\alias{plot_periodicity}
\title{Plot periodicity}
\usage{
plot_periodicity(x, code_name, all_sites.col)
}
\arguments{
\item{x}{an extracted (using \code{\link{extract}}) and flagged (using \code{\link{flag_all}}) table}

\item{code_name}{the code name of interest}

\item{all_sites.col}{the HEX colour pallet for each site}
}
\description{
Plot the periodicity for a CC-HIC event. This is the typical
number of entries for the event per patient per 24 hours.
}
