% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_time}
\alias{plot_time}
\title{Plot Time}
\usage{
plot_time(x, code_name, all_sites.col)
}
\arguments{
\item{x}{an extracted (using \code{\link{extract}}) and flagged (using \code{\link{flag_all}}) table}

\item{code_name}{the code name of interest}

\item{all_sites.col}{the HEX colour pallet for each site}
}
\description{
Plot time values from CC-HIC
}
