% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterise_episodes.R
\name{prep_characterise_episodes}
\alias{prep_characterise_episodes}
\title{Prepare data for characterise episodes}
\usage{
prep_characterise_episodes(connection = NULL, .debug = FALSE)
}
\arguments{
\item{connection}{a connection to the CC-HIC database}
}
\description{
Prepare data for characterise episodes
}
\details{
Extracts data necessary for episode characterisation
}
