% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_timevarying.R
\name{process_episode_timestamp}
\alias{process_episode_timestamp}
\title{Process episode with a timestamp}
\usage{
process_episode_timestamp(df, var_name, metadata)
}
\arguments{
\item{df}{a dataframe containing all episode information (It is unlikely that this
will be accessed directly)}

\item{var_name}{the CC-HIC codename for the current variable being processed}

\item{metadata}{the CC-HIC metadata table}
}
\description{
Process a single episode into a rectangular table with a timestamp column
instead of the usual difftime since admission. The timestamp column corresponds
to the exact timestamp of the event of interest. This is particularly useful for
when you need to combine episodes into spells. Take care around times of clock change.
}
