% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{CC-HIC Data Quality Report}
\usage{
report(database = NULL, username = NULL, password = NULL,
  host = "localhost", port = 5432, sqlite_file = NULL,
  output_folder = NULL, write_plots = TRUE, generate_report = TRUE)
}
\arguments{
\item{database}{the database name for connection}

\item{username}{the database username}

\item{password}{the database password}

\item{host}{the database server host IP address}

\item{port}{the database port number}

\item{sqlite_file}{the filename for the sqlite database (if using it)}

\item{output_folder}{the path name for output files}

\item{write_plots}{true/false to output plots. Otherwise only data processing
takes place, which can be considerably quicker.}

\item{generate_report}{triggers the build of the markdown report}
}
\description{
Produce a data quality report for CC-HIC. This wraps all other functions as
necessary to evaluate the whole database in one sweep. This can take some
time!
}
\details{
Specify the output folder in \code{output_folder}.
}
