% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{report_cases_daily}
\alias{report_cases_daily}
\title{Reports Case Numbers according to day}
\usage{
report_cases_daily(unique_cases_tbl = NULL)
}
\arguments{
\item{unique_cases_tbl}{from \code{pull_cases_all}}
}
\value{
breakdown of unique daily cases
}
\description{
Reports Case Numbers according to day
}
\examples{
report_cases_daily(unique_cases)
}
