% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_perc}
\alias{round_perc}
\title{Round to percentage}
\usage{
round_perc(x, n, digits = 0)
}
\arguments{
\item{x}{numerator}

\item{n}{denominator}

\item{digits}{rounding}
}
\value{
a vector of the same length as \code{x} rounded to the defined digit
  length given as a percentage for display
}
\description{
Round to percentage
}
