% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_score.R
\name{score_episodes}
\alias{score_episodes}
\title{Data Quality Score for CC-HIC Episodes

Creates a simple score based on the number of submitted and number
of validated episodes}
\usage{
score_episodes(connection)
}
\arguments{
\item{connection}{a database connection to the CC-HIC database returned by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}}
}
\value{
a tabulation of episode scores by site
}
\description{
Data Quality Score for CC-HIC Episodes

Creates a simple score based on the number of submitted and number
of validated episodes
}
