% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_score.R
\name{score_events}
\alias{score_events}
\title{Data Quality Score for CC-HIC Events}
\usage{
score_events(connection)
}
\arguments{
\item{connection}{a database connection to the CC-HIC database returned by
[DBI::dbConnect()]}
}
\value{
a tabulation of event scores by site
}
\description{
Data Quality Score for CC-HIC Events
}
