% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_evaluation.R
\name{setup_folders}
\alias{setup_folders}
\title{Setup output folders for data quality evaluation}
\usage{
setup_folders(output_folder = NULL, write_plots = TRUE)
}
\arguments{
\item{output_folder}{}
}
\description{
While some components of the data quality evaluation are stored alongside
the original data, some components are best exported and viewed externally.
This sets up a standard folder structure to recieve these exports.
}
