% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_safely.R
\name{study_safely}
\alias{study_safely}
\title{Study safely}
\usage{
study_safely(ctn, code_names, from, to)
}
\arguments{
\item{code_names}{string vector of code_names to be extracted for the study.
These are referred to as "bellweather" variables for the study. It is
likely that if \emph{all} variables for a study are used here, that no safe
areas will be found. It is better to specify a small number of important
variables, that you believe are indicative of overall quality.}

\item{from}{study starting date}

\item{to}{study ending date}
}
\value{

}
\description{
Study safely
}
\details{
Interrogates the CC-HIC database and provides a vector of episode
ids that can "safely" be used for your study based on the data and date
ranges you require. Think of this as an implementation of a venn diagram
for data quality This means that these episodes meet the following
criteria:
\itemize{
\item Episodes themselves are valid
\item Episodes come from sites that contribute data for the chosen bellweather
fields specified in \code{code_names}.
\item Episodes come from time period where we know data is being contributed.
\item Events are of good quality during the time period examined
}
\subsection{Choosing Bellweather Variables}{

It is better to select a small number of important variables to \code{code_names}.
The more variables that are supplied, the lower the likelihood that any
"safe" data will be found. The best, and most pragmatic, returns from this
function are therefore ones that select a small number (around 10) critical
fields that the study would most certainly not be possible without.
}
}
