% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarise_verification}
\alias{summarise_verification}
\title{Collate and Summarise Verification Information}
\usage{
summarise_verification(verified, stats, coverage, reference)
}
\arguments{
\item{verified}{a verified dataitem from \code{\link{verify_events}}}

\item{stats}{the stats verification or NULL}

\item{coverage}{the coverage verification or NULL}

\item{reference}{the reference table from \code{\link{make_reference}}}
}
\description{
Collate and Summarise Verification Information
}
