% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_main}
\alias{summary_main}
\title{Summary for flagged data item}
\usage{
summary_main(x, reference)
}
\arguments{
\item{x}{a flagged dataframe}
}
\description{
Writes out summary codes following the flagging proceedure
}
\examples{
summary(x)
}
