% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_cchic}
\alias{theme_cchic}
\title{CC-HIC ggplot2 Theme}
\usage{
theme_cchic(...)
}
\arguments{
\item{...}{further arguments to pass to \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\description{
CC-HIC ggplot2 Theme
}
