% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterise_episodes.R
\name{total_unit_admissions}
\alias{total_unit_admissions}
\title{Report total admission numbers for each ICU}
\usage{
total_unit_admissions(events_table = NULL, reference_table = NULL)
}
\arguments{
\item{events_table}{the CC-HIC events database table}

\item{reference_table}{the reference table returned from
\code{\link{make_reference}}}
}
\value{
a tibble with unique episodes and patients reported by ICU
}
\description{
Reports on ICNARC CMP unit codes to describe the numbers of reported cases by
site. This is dependent on accurate parsing of the xml schema for hic code
0002 (INCARC CMP unit code). And as such, may dramatically under-report. To
see a comprehensive list of actual cases that are not dependent upon this,
see \code{\link{weekly_admissions}}.
}
