% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{unit_admissions}
\alias{unit_admissions}
\title{Report Admission Numbers for each Unit}
\usage{
unit_admissions(events_table = NULL, reference_table = NULL)
}
\arguments{
\item{events_table}{the events table}

\item{reference_table}{the reference table}
}
\value{
a tibble with unique episodes and patients reported by ICU
}
\description{
Reports on ICNARC CMP unit codes to describe the numbers of reported cases by site.
This is dependent on accurate parsing of the xml schema for hic code 0002 (INCARC CMP
unit code). And as such, may dramatically under-report. To see a comprehensive list
of acual cases, see the \code{weekly_admissions} function.
}
\examples{
unit_admissions(events, reference)
}
