% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{unit_discharge_status}
\alias{unit_discharge_status}
\title{Collect Unit Discharge Status}
\usage{
unit_discharge_status(event_table)
}
\arguments{
\item{event_table}{main event table}
}
\value{
a tibble with unit discharge status
}
\description{
pulls the discharge status
}
\examples{
unit_discharge_status(events)
}
