% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_cp_01.R
\name{va_cp_01}
\alias{va_cp_01}
\title{Validate VA_CP_01: Missingness of a particular item from the requisite}
\usage{
va_cp_01(tbls)
}
\arguments{
\item{tbls}{}
}
\value{

}
\description{
Validate VA_CP_01: Missingness of a particular item from the requisite
}
