% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_rc_01.R
\name{va_rc_01}
\alias{va_rc_01}
\title{Validate VA_RC_01: Structural data missingness conforms to agreed schema}
\usage{
va_rc_01(tbls)
}
\arguments{
\item{tbls}{}
}
\value{

}
\description{
Validate VA_RC_01: Structural data missingness conforms to agreed schema
}
