% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_episodes}
\alias{validate_episodes}
\title{Validate Episodes Based Upon Patchy Data}
\usage{
validate_episodes(episode_length_table, reference_table, all_sites,
  threshold = 10)
}
\arguments{
\item{threshold}{threshold number of days to use in calculation}

\item{episodes}{collected episodes table}

\item{provenance}{collected provenance table}
}
\value{
a tibble of months that are to be excluded from future analysis
}
\description{
Determines which months are invalid based on a low contribution of data
This is based upon the long term daily average for admissions. Days that
fall below 2 SD of the long term mean are tagged. If more than the
threshold value occur in a single month, the month is removed from
further analysis
}
\details{
This function removes episodes that occur during particularly sparse periods
(as this is likely that these months are contributing poor data) and return
only episodes that have a logical consistency in non-sparse months
}
\examples{
invalid_months(episodes, provenance)
invalid_months(episodes, provenance, theshold = 15)
}
