% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_event}
\alias{validate_event}
\title{Validate CC-HIC Events}
\usage{
validate_event(validated_episodes = NULL, flagged_events = NULL)
}
\arguments{
\item{validated_episodes}{a tibble with 1 column of validated episodes}

\item{flagged_data_item}{a tibble of a CC-HIC event with full condition flagging}
}
\value{
a character vector with event IDs for validated episodes
}
\description{
Validates a CC-HIC event. This should be applied after event flagging has taken place
}
\examples{
validate_field(core, reference, "NIHR_HIC_ICU_0108", qref, episode_length)
}
