% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_codes.R
\name{validate_nhs}
\alias{validate_nhs}
\title{Validate NHS Numbers}
\usage{
validate_nhs(nhs_numbers = NULL)
}
\arguments{
\item{nhs_numbers}{as character vector}
}
\value{
a logical vector
}
\description{
This function will validate nhs numbers according to the checksum process
outlined by UK Government Data Standards Catalogue (GDSC), Version 2.0, Agreed 01.01.02
availible as reference here: https://www.datadictionary.nhs.uk/version2/data_dictionary/data_field_notes/n/nhs_number_de.asp?shownav=0
}
\examples{
validate_nhs("6744424270") # expect TRUE
}
