% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_rc_01.R
\name{ve_rc_01}
\alias{ve_rc_01}
\title{Validate VE_RC_01: Referential integrity between tables}
\usage{
ve_rc_01(tbls)
}
\arguments{
\item{tbls}{}
}
\value{

}
\description{
Validate VE_RC_01: Referential integrity between tables
}
