% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_events.R
\name{verify_bounds}
\alias{verify_bounds}
\title{verify Event Plausibility - Boundary Checks (S3 Generic)}
\usage{
verify_bounds(x, los_table = NULL, hours = 24)
}
\arguments{
\item{x}{an extracted nhic event table}

\item{los_table}{episode length table}

\item{hours}{the number of hours you allow before and after an episode}
}
\value{
a tibble of the same length as x with the following features:
\describe{
  \item{-1}{event is before the ICU episode}
  \item{0}{event is during the ICU episode}
  \item{+1}{event is after the ICU episode}
}
}
\description{
Varifies events as being within (\code{0}), after (\code{+1}) or
before (\code{-1}) an associated episode. This isn't necessarily a problem,
for example, microbiology data coming back after death. However, it is
often demonstrative of a bigger problem. For example, data is sometimes
contributes from more than one episode, and then duplicated across episodes.
}
\examples{
## DB Connection
ctn <- connect(sqlite_file = "./data-raw/synthetic_db.sqlite3")

## Pre-requisites
core <- make_core(ctn)
episodes <- characterise_episodes(ctn)

## Data item extraction
hr <- extract(core, input = "NIHR_HIC_ICU_0108")

## verify Boundary Conditions
vhr <- verify_bounds(hr, episodes)
head(vhr)
DBI::dbDisconnect(ctn)
}
