% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_chronology.R
\name{verify_chronology}
\alias{verify_chronology}
\title{Verify Chronology}
\usage{
verify_chronology(connection)
}
\arguments{
\item{connection}{a database connection}
}
\description{
Verify Chronology
}
