% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_events.R
\name{verify_complete}
\alias{verify_complete}
\title{verify Completeness}
\usage{
verify_complete(x, reference_tbl)
}
\arguments{
\item{x}{an extracted dataitem that has passed through verification}
}
\value{
a tibble with summary information on dataitem completeness
}
\description{
verify Completeness
}
