% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_events.R
\name{verify_coverage}
\alias{verify_coverage}
\title{Verify Event Plausibility - Coverage Checks (S3 Generic)}
\usage{
verify_coverage(x, reference_tbl)
}
\arguments{
\item{x}{an extracted dataitem}

\item{reference_tbl}{reference table from \code{\link{make_reference}}}
}
\value{
a table describing the coverage over a defined time window
}
\description{
Checks to ensure that long term data item contribution is consistent. This is
because often back end changes occur in hospitals that silently disrupt the
ETL process, and as such, some dataitems disappear without warning. There are
three main error prone patterns that this is trying to detect:
- 1. Sudden stopping of data (transitions to zero)
- 2. Sudden change in amount of data (transitions to new lower mean)
- 3. High frequency of anomalous data in general
}
\details{
The return is a table of months and years for which each site is considered to have appropraite
coverage
}
