% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_events.R
\name{verify_duplicate}
\alias{verify_duplicate}
\title{verify Event Plausibility - Duplication Checks (S3 Generic)}
\usage{
verify_duplicate(x, exact = TRUE)
}
\arguments{
\item{x}{an extracted event table}

\item{exact}{TRUE/FALSE}
}
\value{
a tibble of the same length as x with the following features:
\describe{
  \item{+1}{value is a suspected/confirmed duplicate}
  \item{0}{event is unique}
}

Note, when 2 values are deems to be duplicates, the one that first appears
in the database is cleared as verified, while the second is not.
}
\description{
Varifies events as being duplicated (\code{+1}) or not. This is primarily
trained on the datetime of an object (if 2d), and looks for two events that
are perfectly co-incident. This assumptions can be relaxed with
\code{exact = FALSE} which looks to see if there are any duplicates in
value for the preceeding/following 12 hours. This is pointless for dataitems
like heart rate, but is useful so dataitems like creatinine, where we have
seen this type of duplication error. 1d events are checked for raw
duplication, since they should only occur once per episode anyway.
}
\examples{
## DB Connection
ctn <- connect(sqlite_file = "./data-raw/synthetic_db.sqlite3")

## Pre-requisites
core <- make_core(ctn)

## Data item extraction
hr <- extract(core, input = "NIHR_HIC_ICU_0108")

## verify Range
vhr <- verify_duplicate(hr)
head(vhr)
DBI::dbDisconnect(ctn)
}
