% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_episodes.R
\name{verify_event}
\alias{verify_event}
\title{Verify CC-HIC Events}
\usage{
verify_event(validated_episodes = NULL, flagged_events = NULL)
}
\arguments{
\item{validated_episodes}{a tibble with 1 column of validated episodes}

\item{flagged_events}{a tibble of a CC-HIC event with full condition flagging}
}
\value{
a character vector with event IDs for validated episodes
}
\description{
Verfifies a CC-HIC event. This should be applied after event flagging has taken place
}
