% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_events.R
\name{verify_events}
\alias{verify_events}
\title{verify CC-HIC Data}
\usage{
verify_events(x, los_table = NULL)
}
\arguments{
\item{x}{extracted dataitem from \code{\link{extract}}}

\item{los_table}{episode length table from \code{\link{characterise_episodes}}}
}
\value{
a tibble with verification applied
}
\description{
verify CC-HIC Data
}
\details{
Other varification components are found elsewhere, as they don't necessarily
fit into an evaludation at the data item level. I am contemplating how to
unify this procedure.
}
\examples{
## DB Connection
ctn <- connect(sqlite_file = "./data-raw/synthetic_db.sqlite3")

## Pre-requisites
core <- make_core(ctn)
episode_length <- characterise_episodes(ctn)

## Data item extraction
hr <- extract(core, input = "NIHR_HIC_ICU_0108")

## Full varification
vhr <- verify_events(hr, episode_length)
head(vhr)
DBI::dbDisconnect(ctn)
}
