% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_codes.R
\name{verify_icnarc}
\alias{verify_icnarc}
\title{Verify ICNARC Diagnostic Code}
\usage{
verify_icnarc(icnarc_code = NULL)
}
\arguments{
\item{icnarc_code}{as a character vector}
}
\value{
a logical vector
}
\description{
Verifies that a code conforms to the ICNARC Diagnostic code specification.
This is a 5 level hierarchical code that follows a pattern such as
##.##.##.##.##. A code need only supply levels 1:3 to be valid.
}
\details{
Leading zeros in each level are not necessary. Acceptable separators are:
.,|/-
}
\examples{
verify_icnarc("01.06.04") # partial code; still valid
verify_icnarc("01.06.04.20.16") # perfect formatting
verify_icnarc("1.6.4.20.16") # not strictly correct, but can be parsed
verify_icnarc("01-6-04.20.16") # annoying, but can be parsed
verify_icnarc("01-6-04.NA.NA") # will warn about the NAs, but will work.
}
