% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_events.R
\name{verify_periodicity}
\alias{verify_periodicity}
\title{verify Event Plausibility - Periodicity Checks (S3 Generic)}
\usage{
verify_periodicity(x, ...)
}
\arguments{
\item{x}{an extracted events table}

\item{los_table}{episode length table}
}
\value{
x with a periodicity value and validation columns
}
\description{
Provides the periodicity of an extracted data item. This is the number of
submitted varified values per day. As such, the range, boundary and
duplication varification methods must have been first run. Periodicity is
only defined for 2d data.
}
\details{
As periodicity for dataitems is not yet specified, this highlights cases
that are below the group 5th centile \code{-1} and above the 95th centile
\code{+1}

This is a similar concept to the \code{\link{verify_coverage}}, although is
concerned with the episode, rather than site, level.
}
