% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_events.R
\name{verify_periodicity_generic}
\alias{verify_periodicity_generic}
\title{verify periodicity of 2d class}
\usage{
verify_periodicity_generic(x, los_table)
}
\arguments{
\item{x}{extracted dataitem}

\item{los_table}{episode length table}
}
\value{
a mutated tibble from x including a periodicity column
}
\description{
verify periodicity of 2d class
}
