% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_codes.R
\name{verify_post_code}
\alias{verify_post_code}
\title{Varify Post Code}
\usage{
verify_post_code(post_code = NULL)
}
\arguments{
\item{post_code}{character vector of post codes}
}
\value{
a logical vector
}
\description{
Varify that a particular string conforms to a UK Post Code. Note, this does
not actually check a database (i.e. validation) to see if the post code
exists. It only checks that the Post Code meets the correct technical
specification for a postcode.
}
\examples{
varify_post_code("AL5 3HE")
}
