% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_events.R
\name{verify_range.datetime_1d}
\alias{verify_range.datetime_1d}
\title{Range Checks - Datetimes}
\usage{
\method{verify_range}{datetime_1d}(x = NULL)
}
\arguments{
\item{x}{an extracted table}
}
\description{
Verifies that all dates are before now (i.e. not in the future) and not
before 1900-01-01 00:00:00, which seems reasonable.
}
