% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_events.R
\name{verify_range}
\alias{verify_range}
\title{verify Event Plausibility - Range Checks (S3 Generic)}
\usage{
verify_range(x = NULL)
}
\arguments{
\item{x}{an extracted event table}
}
\value{
a tibble of the same length as x with the following features:
\describe{
  \item{-1}{event is below the defined range of plausibility}
  \item{0}{event is valid}
  \item{+1}{event is above the defined range of plausibility}
}
}
\description{
Varifies events as being in-range (\code{0}), high (\code{+1}) or
low (\code{-1}). These ranges have been calibrated based upone; ICNARC
reference ranges, prior evidence (usually case report for exceptional values)
or expert opinion. Reference ranges are all found in \code{\link{qref}}.
}
\examples{
## DB Connection
ctn <- connect(sqlite_file = "./data-raw/synthetic_db.sqlite3")

## Pre-requisites
core <- make_core(ctn)

## Data item extraction
hr <- extract(core, input = "NIHR_HIC_ICU_0108")

## verify Range
vhr <- verify_range(hr)
head(vhr)
DBI::dbDisconnect(ctn)
}
