% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_events.R
\name{verify_range.time_1d}
\alias{verify_range.time_1d}
\title{Range Checks - Times}
\usage{
\method{verify_range}{time_1d}(x = NULL)
}
\arguments{
\item{x}{an extracted table}
}
\description{
Verifies that all times are between 00:00:00 and 23:59:59
}
