% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_events.R
\name{verify_range_numeric}
\alias{verify_range_numeric}
\title{Range Checks - Numeric}
\usage{
verify_range_numeric(x = NULL)
}
\arguments{
\item{x}{an extracted table}
}
\description{
The generic function for numeric type range verification.
}
