% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_events.R
\name{verify_range_string}
\alias{verify_range_string}
\title{Range Checks - String}
\usage{
verify_range_string(x = NULL)
}
\arguments{
\item{x}{an extracted table}
}
\description{
The generic function for string type range verification.
}
