% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_study.R
\name{verify_study}
\alias{verify_study}
\title{Verify Patients based upon Data Item Requirements}
\usage{
verify_study(
  code_names,
  from = NULL,
  to = NULL,
  database = NULL,
  username = NULL,
  password = NULL,
  host = "localhost",
  port = 5432,
  sqlite_file = NULL,
  out_of_bounds = TRUE,
  range_error = TRUE,
  duplicate = TRUE,
  coverage = TRUE
)
}
\value{

}
\description{
Interrogates the CC-HIC database and provides a vector of episode_ids and event_ids that
can "safely" be used for your study based on the dataitems and time restrictions you require. This
can then be plugged into extract_demographics or extract_timevarying to pull out the study data you
require. Each validation setting can be controlled independently.
}
