% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterise_episodes.R
\name{weekly_admissions}
\alias{weekly_admissions}
\title{Report Weekly Admissions}
\usage{
weekly_admissions(reference_table = NULL)
}
\arguments{
\item{reference_table}{returned from \code{make_reference}}
}
\value{
a summary containing patient and episode tallies
}
\description{
Shows the number of patients and unique admissions to each site by week.
}
