% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_notify}
\alias{write_notify}
\title{Write Notify}
\usage{
write_notify(connection, target_name, local_table, verbose = TRUE, ...)
}
\arguments{
\item{connection}{a database connection returned by [DBI::dbConnect()]}

\item{target_name}{character vector length 1 detailing the database table
to write to}

\item{local_table}{the local table object to write out}

\item{verbose}{logical flag to print results to the console}

\item{...}{optional additional arguments to pass to [DBI::dbWriteTable()]}
}
\description{
Write Notify
}
\details{
Writes out local tables into a database, and let's you know when
  it's done. Pass via ... options like: overwrite, append, copy
}
\seealso{
Other connection: 
\code{\link{retrieve_tables}()}
}
\concept{connection}
