% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{xml_stats}
\alias{xml_stats}
\title{XML Parsing Errors}
\usage{
xml_stats(importstats)
}
\arguments{
\item{importstats}{uncollected importstats postGres table}
}
\value{
a tibble containing summary of XML errors
}
\description{
Interrogates and summarises XML parsing errors
}
\examples{
xml_stats(tbls[["importstats"]])
}
