% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{xml_validation}
\alias{xml_validation}
\title{Checks validation for XML files and returns a summary}
\usage{
xml_validation(xml_file, schema)
}
\arguments{
\item{xml_file}{the xml file you wish to validate}

\item{schema}{the xml schema against which you are validating}
}
\value{
a character vector of linting errors
}
\description{
Checks validation for XML files and returns a summary
}
\examples{
xml_validation(xml, schema)
}
