% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_heat.R
\name{autoplot.heat_cal}
\alias{autoplot.heat_cal}
\title{Plot Calendar Heatmap}
\usage{
autoplot.heat_cal(object, ...)
}
\arguments{
\item{object}{a heat_cal object}
}
\description{
Builds the ggplot layers required to correctly display the calendar heatmap
}
