% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.integer_2d}
\alias{autoplot.integer_2d}
\title{Plot Integer}
\usage{
\method{autoplot}{integer_2d}(x, sites_col = NULL, verify = TRUE)
}
\arguments{
\item{x}{a varified data frame}

\item{sites_col}{the HEX colour pallet for each site}
}
\description{
Plot integer values from CC-HIC
}
