% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean Table}
\usage{
clean(tbl = NULL, dq_ref = NULL)
}
\arguments{
\item{tbl}{a table extracted from the CC-HIC database}

\item{dq_ref}{a tibble if you want to override the default boundary
conditions with the following column headings:
\describe{
  \item{code_name}{the code name for the data item as it appears in your
    table}
  \item{range_min}{the lower bound you want to set}
  \item{range_max}{the upper bound you want to set}
}}
}
\value{
a table of the same dimentions as \code{tbl} but with values cleaned
}
\description{
This function takes an extracted table from either
\code{\link{extract_demographics}} or \code{\link{extract_timevarying}} and
returns the table having dealt with out of range values. Currently this
will either modify the value to NA, or limit the value at an appropriate max
or min.
}
