% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect}
\alias{connect}
\title{Establishes a connection to the database}
\usage{
connect(database = NULL, host = "localhost", port = 5432,
  username = NULL, password = NULL, sqlite_file = NULL)
}
\arguments{
\item{database}{target db}

\item{host}{db host: typically "localhost"}

\item{username}{db username}

\item{password}{db password}

\item{system}{choice of either "postgres" or "sqlite" depending upon your backend}

\item{file}{filename as character vector if using an sqlite db}
}
\value{
a database object connection
}
\description{
Establishes a connection to the database
}
\examples{
ctn <- connect(sqlite_file = "path/to/file.sqlite3")
ctn <- connect(database = "hic", username = "abdce", password = "qwerty")
}
