% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect}
\alias{connect}
\title{Establishes a connection to the database}
\usage{
connect(database = NULL, host = "localhost", port = 5432,
  username = NULL, password = NULL, sqlite_file = NULL)
}
\arguments{
\item{database}{target db}

\item{host}{db host: typically "localhost"}

\item{port}{db port number}

\item{username}{db username}

\item{password}{db password}

\item{sqlite_file}{filename as character vector if using an sqlite db (in
which case this is the only argument that needs to be supplied)}
}
\value{
a database object connection
}
\description{
Establishes a connection to the database
}
\examples{
db_pth <- system.file("testdata/synthetic_db.sqlite3", package = "inspectEHR")
ctn <- connect(sqlite_file = db_pth)
class(ctn)
DBI::dbDisconnect(ctn)
}
