% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Describes the coverage for reporting of a variable (S3 Generic)}
\usage{
coverage(x, ...)
}
\arguments{
\item{x}{an extracted nhic dataitem}

\item{...}{}
}
\value{
a tibble describing the coverage of that dataitem over a defined time window
}
\description{
Describes the coverage for reporting of a variable (S3 Generic)
}
\examples{
coverage(x, occupancy, all_cases)
}
