% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarHeat.R, R/calendar_heat.R
\name{create_calendar_template}
\alias{create_calendar_template}
\title{Create a Calendar Template for Count Data}
\usage{
create_calendar_template(x = NULL, start_date = "2014-01-01",
  end_date = "2018-01-01")

create_calendar_template(x = NULL, start_date = "2014-01-01",
  end_date = "2018-01-01")
}
\arguments{
\item{x}{count data to be used in the calendar
\describe{
  \item{date}{date column formatted to a POSIX date}
  \item{count_column}{counting column of integer type}
}}

\item{start_date}{a starting date as character vector of format YYYY/MM/DD}

\item{end_date}{an end date as character vector of format YYYY/MM/DD}

\item{start_date}{a starting date as character vector of format YYYY/MM/DD}

\item{end_date}{an end date as character vector of format YYYY/MM/DD}

\item{x}{count data to be used in the calendar
\describe{
  \item{date}{date column formatted to a POSIX date}
  \item{count_column}{counting column of integer type}
}}
}
\value{
a tibble with correct week alignments for a calendar heatmap

a tibble with correct week alignments for a calendar heatmap
}
\description{
Creates the basic template for a heatmap calendar in ggplot
when using count data. The calendar will automatically expand
the date contrains to full years

Creates the basic template for a heatmap calendar in ggplot
when using count data. The calendar will automatically expand
the date contrains to full years
}
\examples{
sample_data <- create_calendar_template(myData, "2014-01-01", "2018-01-01")
sample_data <- create_calendar_template(myData, "2014-01-01", "2018-01-01")
}
