% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarHeat.R, R/calendar_heat.R
\name{create_grid}
\alias{create_grid}
\title{Create Calendar Grid}
\usage{
create_grid(calendar_template = NULL)

create_grid(calendar_template = NULL)
}
\arguments{
\item{calendar_template}{a calendar template from \code{create_calendar_template}}

\item{calendar_template}{a calendar template from \code{create_calendar_template}}
}
\value{
a grid for assembling a ggheatcal

a grid for assembling a ggheatcal
}
\description{
This calulates the correct gridlines that need to be drawn when formatting
the calendar.

This calulates the correct gridlines that need to be drawn when formatting
the calendar.
}
\examples{
create_grid(cal_template)
create_grid(cal_template)
}
