% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract NHIC Events}
\usage{
extract(core_table = NULL, input = NULL)
}
\arguments{
\item{core_table}{core table from make_core}

\item{input}{the input variable of choice}
}
\value{
A tibble with 1 row per event
}
\description{
Extracts NHIC events and appends them with the correct class for
further processing. This is essentially the creator method for
the S3 classes associated with the inspectEHR package
}
\examples{
db_pth <- system.file("testdata/synthetic_db.sqlite3", package = "inspectEHR")
ctn <- connect(sqlite_file = db_pth)
core <- make_core(ctn)
extract(core, input = "NIHR_HIC_ICU_0073")
DBI::dbDisconnect(ctn)
}
