% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract NHIC Events}
\usage{
extract(core_table = NULL, input = "NIHR_HIC_ICU_0557")
}
\arguments{
\item{core_table}{core table from make_core}

\item{input}{the input variable of choice}
}
\value{
A tibble with 1 row per event
}
\description{
Extracts NHIC events and appends them with the correct class for
further processing. This is essentially the creator method for
the S3 classes associated with the inspectEHR package
}
\examples{
extract(core)
extract(core, input = "NIHR_HIC_ICU_0557")
}
