% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_demographics.R
\name{extract_demographics}
\alias{extract_demographics}
\title{Extract 1d data from CC-HIC}
\usage{
extract_demographics(events = NULL, metadata = NULL,
  code_names = "NIHR_HIC_ICU_0093", rename = NULL)
}
\arguments{
\item{events}{a database events table}

\item{metadata}{a database metadata table}

\item{code_names}{a character vector of HIC codes you want to retrieve}

\item{rename}{a character vector of names you want to relabel column names
as, or NULL ( the default) if you want column names to be labelled with
HIC codes}
}
\value{
A tibble of 1d data
}
\description{
This function takes remote database tables from the CC-HIC project and a
vector of HIC codes and returns a table with 1 row per patient and 1 column
per data item.
}
\examples{
extract_demographics(tbls[["variables"]], tbls[["events"]])
}
