% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_timevarying.R
\name{extract_timevarying}
\alias{extract_timevarying}
\title{Reshape timevarying events into a long form timeseries table}
\usage{
extract_timevarying(events, metadata, code_names, chunk_size = 5000,
  cadance = 1, rename = NULL)
}
\arguments{
\item{events}{database events table}

\item{metadata}{database metadata table (collected)}

\item{code_names}{a vector of HIC codes names to be wrangled out}

\item{chunk_size}{a chunking parameter to help speed up the function and manage memory constaints}

\item{cadance}{a numerical scalar to describe the base time unit to build
each row, in divisions of an hour. For exampple: 1 = 1 hour, 0.5 = 30 mins, 2 = 2 hourly.
If you set cadance = "exact", then the EXACT datetime will be used at the time column.
This is likely to generate a LARGE table, so use cautiously.}
}
\value{
sparse tibble with hourly cadance as rows, and unique hic events as columns
}
\description{
This is the workhorse function that transcribes 2d data from the HIC DB
to a dense table with 1 column per variable (and metadata if relevent)
and 1 row per time unit per patient.
}
\details{
The time unit is user definable, and set by the "cadance" argument.
The default behaviour is to procude a table with 1 row per hour per patient.

Many events inside CC-HIC occur on a greater than hourly basis. Depending upon
the chosen analysis, you may which to increase the cadance. 0.5 for example will produce
a table with 1 row per 30 minutes per patient.

Choose what variables you want to pull out wisely. Whilst this is actually
pretty quick considering what it needs to do, it can take a very long time (24 hours or more)

It is perfectly possible for this table to produce negative time rows. If, for example
a patient had a measure taken in the hours before they were admitted, then this would
be added to the table with a negative time value. As a concrete example,
if a patient had a sodium measured at 08:00, and they were admitted to the
ICU at 20:00 the same day, then the sodium would present on the output table
at time = -12. This is normal behaviour it is left to the end user to
descibe how best they wish to account for this.
}
\examples{
\dontrun{
extract_timevarying(tbls[["events"]], collect(tbls[["variables"]]), "NIHR_HIC_ICU_0108")
}
}
