% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_heat.R
\name{find_first_sunday}
\alias{find_first_sunday}
\title{Find First Sunday}
\usage{
find_first_sunday(x)
}
\arguments{
\item{x}{a vector of class date}
}
\value{
a vector with dates for the first sunday of each year supplied to x
}
\description{
Finds the first sunday of the month supplied to x
}
