% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{identify_spells}
\alias{identify_spells}
\title{Indentify Spells}
\usage{
identify_spells(episode_length = NULL, episodes = NULL, minutes = 60)
}
\arguments{
\item{episode_length}{episode length tibble}

\item{episodes}{episodes tibble}

\item{minutes}{numeric value to define transition period}
}
\value{
a tibble with spells identified and coded
}
\description{
some sites have patients check out of one ICU and into another (for example ICU stepdown to HDU).
This checks to see if patients are discharged from 1 unit and admitted to another wihtin a pre-defined
time period, specified in the minutes argument.
}
\details{
This only evaluates episodes that have already been flagged as valid by the episode_length function.
}
\examples{
identify_spells(episode_length, episodes)
}
