% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-inspectEHR.R
\docType{package}
\name{inspectEHR}
\alias{inspectEHR}
\title{inspectEHR}
\description{
This package contains functions for data wrangling, cleaning, verification,
validation and reporting. It is connected to \pkg{cleanEHR}, where eventually
we plan to migrate the wrangling and cleaning components.
}
