% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ks_test}
\alias{ks_test}
\title{Pairwise Kolmogorov-Smirnov Distance}
\usage{
ks_test(x)
}
\arguments{
\item{x}{extracted data item from \code{\link{extract}}}
}
\description{
Takes a distribution from a CC-HIC dataitem and returns
the pairwise KS distance by site.
}
