% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_hic_code.R
\name{lookup_hic}
\alias{lookup_hic}
\title{Lookup CC-HIC Code}
\usage{
lookup_hic(search = 108)
}
\arguments{
\item{search}{search term as either a code (numeric) or name (character)}
}
\value{
a table with query results
}
\description{
Perform a quick search for a CC-HIC Code
}
\examples{
lookup_hic()
}
