% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_nihr_code.R
\name{lookup_hic}
\alias{lookup_hic}
\title{Lookup HIC Value}
\usage{
lookup_hic(search = 108)
}
\arguments{
\item{search}{search term as either a code (numeric) or name (character)}
}
\value{
a table with query results
}
\description{
Performs a quick search of the HIC data catalogue
}
\examples{
lookup_hic()
}
