% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_core}
\alias{make_core}
\title{Core table}
\usage{
make_core(connection)
}
\arguments{
\item{connection}{a database connection}
}
\value{
a database object
}
\description{
Writes a remote SQL query to make a core table with all the necessary column
names for most data extraction tasks. No work is done till the function is
explicitly called.
}
\examples{
db_pth <- system.file("testdata/synthetic_db.sqlite3", package = "inspectEHR")
ctn <- connect(sqlite_file = db_pth)
core <- make_core(ctn)
head(core)
DBI::dbDisconnect(ctn)
}
