% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_core}
\alias{make_core}
\title{Make core table}
\usage{
make_core(connection)
}
\arguments{
\item{connection}{a database connection}
}
\value{
a tibble with all hic events
}
\description{
Produces the core table structure necessary for the data quality report
}
\examples{
make_core(ctn)
}
