% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{make_dict}
\alias{make_dict}
\title{Determine dimentionality of Events}
\usage{
make_dict(metadata = NULL)
}
\arguments{
\item{metadata}{a collected metadata table}
}
\value{
a metadata table
}
\description{
This function take a metadata table and produces a list of NIHR codes
describing whether the data is 1d (i.e. only measured once) or 2d (i.e "time
varying")
}
\examples{
makeDict(collect(tbls[["variables"]]))
}
