% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_heat.R
\name{make_heatcal}
\alias{make_heatcal}
\title{Make a heat_cal object}
\usage{
make_heatcal(reference_table = NULL, type = NULL, site = NULL,
  start_date = "2014-01-01", end_date = "2019-01-01", max_limit = 25)
}
\arguments{
\item{reference_table}{reference table produced from \code{\link{make_reference}}}

\item{site}{character string of hospital site to plot}

\item{start_date}{a starting date as character vector of format YYYY-MM-DD}

\item{end_date}{an end date as character vector of format YYYY-MM-DD}

\item{max_limit}{the upper bound for determining colour scale. This is useful
if you have extreme outliers to enable good constrast over the range of
interest}
}
\value{
an object of class heat_cal
}
\description{
Make a heat_cal object
}
\examples{
\dontrun{
make_heatcal(ref, "UCL")
}
}
