% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_reference}
\alias{make_reference}
\title{Make reference Table}
\usage{
make_reference(connection)
}
\arguments{
\item{connection}{a database connection}
}
\value{
a tibble with episode level data with site
}
\description{
Makes the reference table used my many functions in this package.
This is part of the basic setup.
}
\examples{
db_pth <- system.file("testdata/synthetic_db.sqlite3", package = "inspectEHR")
ctn <- connect(sqlite_file = db_pth)
ref <- make_reference(ctn)
head(ref)
DBI::dbDisconnect(ctn)
}
