% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ecdf}
\alias{plot_ecdf}
\title{Plot CC-HIC Data item eCDF}
\usage{
plot_ecdf(x, sites_col = NULL, verify = TRUE)
}
\arguments{
\item{x}{a varified data frame}

\item{sites_col}{the HEX colour pallet for each site}

\item{verify}{logical indicator if only verified values should be used}
}
\description{
Plot CC-HIC Data item eCDF
}
