% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot Histogram}
\usage{
plot_hist(x, sites_col = NULL, verify = TRUE)
}
\arguments{
\item{x}{an extracted hic data item}

\item{sites_col}{the HEX colour pallet for each site}

\item{verify}{\code{logical} if use only verified data}
}
\description{
Plot a histogram of a CC-HIC event, with each site normalised to 100%
}
