% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ks}
\alias{plot_ks}
\title{Plot Kolmogorov-Smirnov Test}
\usage{
plot_ks(x, reference_tbl)
}
\arguments{
\item{x}{an object from \code{ks_test}}
}
\description{
Produces a plot of the KS distances between sites.
This is quite hard coded at the moment, and so could be improved in a future version
}
