% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_periodicity}
\alias{plot_periodicity}
\title{Plot periodicity}
\usage{
plot_periodicity(x, sites_col = NULL, verify = TRUE)
}
\arguments{
\item{x}{and extracted and verified dataitem}

\item{sites_col}{the HEX colour pallet for each site}
}
\description{
Plot the periodicity for a CC-HIC event. This is the typical
number of entries for the event per patient per 24 hours. Please note,
this can ONLY plot verified events, as the periodicity calculation is
conditional on there being a well defined episode start and end.
}
