% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{qref}
\alias{qref}
\title{QREF - Data Quality reference file for CC-HIC. Version 1}
\format{A tibble with 255 rows and 11 variables:
\describe{
  \item{code_name}{}
  \item{short_name}{}
  \item{long_name}{}
  \item{primary_column}{}
  \item{type}{}
  \item{class}{}
  \item{range_min}{}
  \item{range_max}{}
  \item{ref_lower}{}
  \item{ref_upper}{}
  \item{assumed_units}{}
}}
\usage{
qref
}
\description{
A dataset containing useful metadata of each event within CC-HIC
}
\keyword{datasets}
