% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieve_tables}
\alias{retrieve_tables}
\title{Retrieve Database Tables}
\usage{
retrieve_tables(connection, schema)
}
\arguments{
\item{connection}{an sql connection}
}
\value{
a list containing pointers to tables within the sql connection.
}
\description{
Places all tables from the database connection in a list. This makes use of the tables
in dplyr extremely easy.
}
