% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieve_tables}
\alias{retrieve_tables}
\title{Retrieves db tables}
\usage{
retrieve_tables(connection)
}
\arguments{
\item{connection}{an sql connection}
}
\value{
a list containing pointers to tables within the sql connection
}
\description{
These tables come as a list and must be unlisted or accessed directly
}
\examples{
retrieve_tables(ctn)
}
