% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_any}
\alias{round_any}
\title{Round any}
\usage{
round_any(x, accuracy = 1)
}
\arguments{
\item{accuracy}{}
}
\description{
rounds a numeric value to any arbitrary degree of precision.
defaults to nearest whole integer
}
