% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarise_verification}
\alias{summarise_verification}
\title{Collate and Summarise Verification Information}
\usage{
summarise_verification(x, reference)
}
\arguments{
\item{x}{a verified dataitem from \code{\link{verify_events}}}

\item{reference}{the reference table from \code{\link{make_reference}}}
}
\description{
Collate and Summarise Verification Information
}
\examples{
summary(x)
}
