% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_events.R
\name{verify_events}
\alias{verify_events}
\title{Verify CC-HIC Data}
\usage{
verify_events(x, los_table = NULL)
}
\arguments{
\item{x}{extracted dataitem from \code{\link{extract}}}

\item{los_table}{episode length table from \code{\link{characterise_episodes}}}
}
\value{
a tibble with verification applied
}
\description{
Applies all relevent varification flags to an extracted dataitem. This
includes:
\itemize{
  \item Completeness: contributed data items match local capability (i.e.
    missingness only occurs when the data doesn't exist) Performed by
    \code{\link{verify_complete}}
  \item Uniqueness plausibility: descriptions of singular events/objects are
    not duplicated. Performed by \code{\link{verify_unique_plausible}}
  \item Atemporal plausibility: Events occur within their episode (within a
    reasonable buffer time). Events fall within an accepted range, follow the
    expected distribution and agree with internal/local knowledge. Repeated
    measures of the same event show the expected variability. Performed by
    \code{\link{verify_atemporal_plausible}}
  \item Temporal plausibility: value density over time are consistent with
    local expectations
    }
}
\details{
Other varification components are found elsewhere, as they don't necessarily
fit into an evaludation at the data item level. I am contemplating how to
unify this procedure.
}
\examples{
## DB Connection
db_pth <- system.file("testdata/synthetic_db.sqlite3", package = "inspectEHR")
ctn <- connect(sqlite_file = db_pth)

## Pre-requisites
core <- make_core(ctn)
episode_length <- characterise_episodes(ctn)

## Data item extraction
hr <- extract(core, input = "NIHR_HIC_ICU_0108")

## Full varification
vhr <- verify_events(hr, episode_length)
head(vhr)
DBI::dbDisconnect(ctn)
}
