% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_codes.R
\name{verify_icnarc}
\alias{verify_icnarc}
\title{Varify ICNARC Diagnostic Code}
\usage{
verify_icnarc(icnarc_code = NULL)
}
\arguments{
\item{icnarc_code}{as a character vector}
}
\value{
a logical vector
}
\description{
Varifies that a code conforms to the ICNARC Diagnostic code specification.
This is a 5 level hierarchial code that follows a pattern such as
##.##.##.##.##.
}
\details{
Leading zeros in each level are not necessary. Acceptable separators are:
.,|/-
}
\examples{
verify_icnarc("01.06.04.20.16") # correctly formatted
verify_icnarc("1.6.4.20.16") # not strictly correct, but parsable
verify_icnarc("01-6-04.20.16") # annoying, but parsable
}
