% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_parameters.R
\name{weekly_admissions}
\alias{weekly_admissions}
\title{Report Weekly Admissions}
\usage{
weekly_admissions(distinct_episodes = NULL)
}
\arguments{
\item{distinct_episodes}{from reference table}
}
\value{
a summary containing patient and episode numbers
}
\description{
Shows the number of admissions to each each in terms of
patients and episodes stratified in weeks
}
\examples{
weekly_admissions(distinct_episodes)
}
